split the scripts out individually and other minor fixes
